requirements.txt
astroid==2.3.3
click==7.1.1
dnspython==1.16.0
eventlet==0.25.2
fake-rpi==0.7.0
Flask==1.1.2
Flask-SocketIO==4.2.1
greenlet==0.4.15
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
numpy==1.18.3
pylint==2.4.4
python-engineio==3.12.1
python-socketio==4.5.1
RPi.GPIO==0.7.0
six==1.14.0
Werkzeug==1.0.1
wrapt==1.11.2
